feat: z-index 상수화 및 오버레이 컴포넌트 구조 개선
- z-index 값을 상수로 관리하도록 변경
- 오버레이 컴포넌트들의 렌더링 순서 최적화
- 버전 0.1.1로 업데이트
- 의존성 패키지 업데이트
