+++ example/app/_layout.tsx
@@ -1,16 +1,20 @@
+import { Stack } from 'expo-router';
 import { ThemeProvider, OverlayProvider } from 'zs-ui';
-import NavigateComponent from './src/NavigateComponent';
-import SnackBar from './src/ui/Snackbar';
 import { SafeAreaProvider } from 'react-native-safe-area-context';
+import SnackBar from '../src/ui/Snackbar';
 
-export default function App() {
+export default function RootLayout() {
   return (
     <SafeAreaProvider>
       <ThemeProvider>
         <OverlayProvider customSnackbar={SnackBar}>
-          <NavigateComponent />
+          <Stack
+            screenOptions={{
+              animation: 'slide_from_right',
+            }}
+          />
         </OverlayProvider>
       </ThemeProvider>
     </SafeAreaProvider>
   );
-}
+}
