+++ example/package.json
@@ -1,7 +1,7 @@
 {
   "name": "zs-ui-example",
   "version": "1.0.0",
-  "main": "expo/AppEntry.js",
+  "main": "expo-router/entry",
   "scripts": {
     "start": "expo start",
     "android": "expo run:android",
@@ -12,11 +12,16 @@
     "@react-native-async-storage/async-storage": "1.23.1",
     "expo": "52.0.46",
     "expo-blur": "~14.0.3",
+    "expo-constants": "~17.0.8",
+    "expo-linking": "~7.0.5",
     "expo-navigation-bar": "~4.0.7",
+    "expo-router": "~4.0.20",
+    "expo-status-bar": "~2.0.1",
     "react": "18.3.1",
     "react-native": "0.76.9",
     "react-native-reanimated": "~3.17.4",
     "react-native-safe-area-context": "4.12.0",
+    "react-native-screens": "~4.4.0",
     "react-native-svg": "15.8.0"
   },
   "devDependencies": {
