+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@0610studio/zs-ui",
-  "version": "0.1.0",
+  "version": "0.1.1",
   "private": false,
   "description": "EXPO ZS-UI",
   "type": "module",
